allow monitoring all queues, not just fulltext
